Security fix

+++ inc/user.class.php
@@ -1273,16 +1273,16 @@ class User extends CommonDBTM {
       $userUpdated = false;
 
       if (isset($this->input['_useremails']) && count($this->input['_useremails'])) {
-         $useremail = new UserEmail();
          foreach ($this->input['_useremails'] as $id => $email) {
             $email = trim($email);
 
             // existing email
-            if ($id > 0) {
+            $useremail = new UserEmail();
+            if ($id > 0 && $useremail->getFromDB($id) && $useremail->fields['users_id'] === $this->getID()) {
                $params = ['id' => $id];
 
-               // empty email : delete
-               if (strlen($email) == 0) {
+               if (strlen($email) === 0) {
+                  // Empty email, delete it
                   $deleted = $useremail->delete($params);
                   $userUpdated = $userUpdated || $deleted;
 
@@ -1292,7 +1292,8 @@ class User extends CommonDBTM {
 
                   $existingUserEmail = new UserEmail();
                   $existingUserEmail->getFromDB($id);
-                  if ($params['email'] == $existingUserEmail->fields['email']
+                  if ($existingUserEmail->getFromDB($id)
+                      && $params['email'] == $existingUserEmail->fields['email']
                       && $params['is_default'] == $existingUserEmail->fields['is_default']) {
                      // Do not update if email has not changed
                      continue;
@@ -1302,9 +1303,12 @@ class User extends CommonDBTM {
                   $userUpdated = $userUpdated || $updated;
                }
 
-            } else { // New email
-               $email_input = ['email'    => $email,
-                               'users_id' => $this->fields['id']];
+            } else {
+               // New email
+               $email_input = [
+                  'email'    => $email,
+                  'users_id' => $this->fields['id']
+               ];
                if (isset($this->input['_default_email'])
                    && ($this->input['_default_email'] == $id)) {
                   $email_input['is_default'] = 1;
@@ -2436,7 +2440,7 @@ JAVASCRIPT;
             echo "</td></tr>";
          }
 
-         if ($this->can($ID, UPDATE)) {
+         if ($caneditpassword) {
             echo "<tr class='tab_bg_1'><th colspan='4'>". __('Remote access keys') ."</th></tr>";
 
             echo "<tr class='tab_bg_1'><td>";
